Update 05-discussion.Rmd
In line 8 what do you mean by "in those with mental stress" - referring to depression/PTSD? Or are you referring to stress-induced HRV?
For the paragraph starting in line 27, consider the differences between the twins cohort and MIMS/MIPS cohort as well. Could the relationship of depression and HRV be attenuated in those with CAD because of competing risk? Also consider the difference in using lifetime depression (SCID) vs. current depressive symptoms...lifetime depression less likely to show and effect because HRV is more reflective of the current state rather than history of MDD